hidden_objectives_code_sample/requirements.txt
# Core ML/DL dependencies
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.15.0

# Evaluation and metrics
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Experiment tracking
wandb>=0.16.0
tqdm>=4.66.0

# Utilities
pyyaml>=6.0
jsonlines>=4.0.0
pandas>=2.1.0

# For activation caching and analysis
einops>=0.7.0
jaxtyping>=0.2.25

# Testing
pytest>=7.4.0

